refactor: Navigator(라우팅 객체), DepengencyManager(의존성 관리 객체) 분리
